pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dialekt-manifest-validator"
version = "0.1.0"
description = "Validator for dialekt agent manifests (spec v1.0.1)"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["dialekt", "manifest", "validator", "agent"]
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "semver>=3.0",
    "croniter>=2.0",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
dialekt-validate-manifest = "dialekt_manifest.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dialekt_manifest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
